+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   'mlsum-es' dataset.
 
 ### Added
-- Added support for Finnish 🇫🇮! This includes the binary sentiment classification dataset [ScandiSent](https://github.com/timpal0l/ScandiSent)
+- Added support for Finnish 🇫🇮! This includes the Finnish part of the reading comprehension dataset
+  [TydiQA-fi](https://huggingface.co/datasets/google-research-datasets/tydiqa/viewer/secondary_task?views%5B%5D=secondary_task_train),
+  the Finnish part of the binary sentiment classification dataset [ScandiSent](https://github.com/timpal0l/ScandiSent)
 
 ### Fixed
 - Evaluating a specific model revision did not work for adapter models, as there was a